+++ src/views/HomeView.vue
@@ -6,7 +6,7 @@ import HomeWeatherDaysSelect from '@/components/home/HomeWeatherDaysSelect.vue'
 import { Location } from '@/domain/entities/Location';
 import { formatedDate } from "@/utils/dateHelper"
 
-const { weatherData, getWeatherData } = useOpenMeteo();
+const { weatherData, getWeatherData, loading, error } = useOpenMeteo();
 
 // i am using lisbon as default
 const latitude = ref(38.713);
@@ -14,11 +14,11 @@ const longitude = ref(-9.139);
 const startDate = ref<Date>(formatedDate(new Date()))
 const endDate = ref<Date>(formatedDate(new Date()))
 const showForecastedDate = ref<boolean>(true)
-const durationInDays = ref<number | null>(null)
+const durationInDays = ref<number>(7)
 
 const fetchLocationData = () => {
   const locationPointer = Location.create(latitude.value, longitude.value)
-  getWeatherData(locationPointer)
+  getWeatherData(locationPointer, undefined, durationInDays.value, undefined, undefined)
 }
 
 const getUserCurrentLocation = () => {
@@ -31,6 +31,7 @@ const getUserCurrentLocation = () => {
     },
     (err) => {
       // TODO: error handling here
+      // snackbar can be implemented to show timed errors
     }
   );
 }
@@ -42,7 +43,7 @@ watch([startDate, endDate], () => {
 
 watch(durationInDays, () => {
   const locationPointer = Location.create(latitude.value, longitude.value)
-  getWeatherData(locationPointer, undefined, parseInt(durationInDays.value) || 7, undefined, undefined)
+  getWeatherData(locationPointer, undefined, durationInDays.value || 7, undefined, undefined)
 })
 
 fetchLocationData()
@@ -52,13 +53,15 @@ fetchLocationData()
 
 <template>
   <main class="flex flex-col gap-2 bg-gray-50 rounded">
+    {{ currentLocationError }}
     <div class="flex items-center p-2 gap-2">
       <input class="bg-white px-3 py-2 w-48 max-w-full" placeholder="latitude" id="latitude" v-model.number="latitude"
         type="number" min="-90" max="90" step="0.0001" />
 
       <input class="bg-white px-3 py-2 w-48 max-w-full" placeholder="longitude" id="longitude" v-model.number="longitude"
         type="number" min="-180" max="180" step="0.0001" />
-      <button class="bg-gray-300 px-3 py-2 rounded cursor-pointer" @click="fetchLocationData">Search</button>
+      <button class="bg-gray-300 px-3 py-2 rounded cursor-pointer disabled:text-gray-100" @click="fetchLocationData"
+        :disabled="!latitude || !longitude">Search</button>
       <button class="bg-white px-3 py-2 cursor-pointer" @click="getUserCurrentLocation">
         use current location
       </button>
@@ -77,6 +80,9 @@ fetchLocationData()
       <HomeWeatherDaysSelect v-else v-model="durationInDays" />
 
     </div>
+
+    <div v-if="loading">fetching...</div>
+    <div v-if="error" class="p-2 text-red-500">{{ error }}</div>
 
     <div v-if="weatherData.length > 0">
       <HomeWeatherTable :weather-data="weatherData" />