+++ CHANGELOG.md
@@ -8,3 +8,4 @@
     2. Add table layout
     3. Show More in case more then 1 day of data
 5. Show historical weather data with start and end date selection
+6. Add error handling when not valid request to open meteo